+++ src/main/java/com/nhnacademy/boostorenginx/controller/BirthdayCouponController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.boostorenginx.controller;
 
 import com.nhnacademy.boostorenginx.dto.birthday.BirthdayCouponRequestDto;
-import com.nhnacademy.boostorenginx.service.BirthdayCouponService;
+import com.nhnacademy.boostorenginx.service.impl.BirthdayCouponServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,12 +14,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class BirthdayCouponController {
 
-    private final BirthdayCouponService birthdayCouponService;
+    private final BirthdayCouponServiceImpl birthdayCouponServiceImpl;
 
     // 생일쿠폰 발급 API
     @PostMapping("/birthday")
     public ResponseEntity<String> issueBirthdayCoupon(@RequestBody BirthdayCouponRequestDto requestDto) {
-        birthdayCouponService.issueBirthdayCoupon(requestDto);
+        birthdayCouponServiceImpl.issueBirthdayCoupon(requestDto);
         return ResponseEntity.ok("생일 쿠폰이 성공적으로 발급되었습니다");
     }
 }
