+++ src/main/java/com/nhnacademy/boostorenginx/controller/CategoryCouponController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.boostorenginx.controller;
 
 import com.nhnacademy.boostorenginx.dto.couponpolicy.CouponPolicySaveRequestDto;
-import com.nhnacademy.boostorenginx.service.CategoryCouponService;
+import com.nhnacademy.boostorenginx.service.impl.CategoryCouponServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,12 +11,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/coupons")
 @RestController
 public class CategoryCouponController {
-    private final CategoryCouponService categoryCouponService;
+    private final CategoryCouponServiceImpl categoryCouponServiceImpl;
 
     // 카테고리 쿠폰 생성
     @PostMapping("/category/{keyword}")
     public ResponseEntity<String> categoryCoupon(@PathVariable("keyword") String keyword, @RequestBody CouponPolicySaveRequestDto couponPolicySaveRequestDto) {
-        categoryCouponService.issueCategoryCoupon(keyword, couponPolicySaveRequestDto);
+        categoryCouponServiceImpl.issueCategoryCoupon(keyword, couponPolicySaveRequestDto);
         return ResponseEntity.ok("카테고리 쿠폰이 성공적으로 발급되었습니다");
     }
