+++ src/main/java/com/nhnacademy/boostorenginx/controller/SellingBookCouponController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.boostorenginx.controller;
 
 import com.nhnacademy.boostorenginx.dto.couponpolicy.CouponPolicySaveRequestDto;
-import com.nhnacademy.boostorenginx.service.SellingBookCouponService;
+import com.nhnacademy.boostorenginx.service.impl.SellingBookCouponServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,11 +11,11 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/coupons")
 @RestController
 public class SellingBookCouponController {
-    private final SellingBookCouponService sellingBookCouponService;
+    private final SellingBookCouponServiceImpl sellingBookCouponServiceImpl;
 
     @PostMapping("/selling-books/{searchKeyword}")
     public ResponseEntity<String> createSellingBookCoupon(@PathVariable String searchKeyword, @RequestBody CouponPolicySaveRequestDto couponPolicySaveRequestDto) {
-        sellingBookCouponService.createCouponForSellingBook(searchKeyword, couponPolicySaveRequestDto);
+        sellingBookCouponServiceImpl.createCouponForSellingBook(searchKeyword, couponPolicySaveRequestDto);
         return ResponseEntity.status(HttpStatus.CREATED).body("도서 쿠폰이 성공적으로 발급되었습니다");
     }
 }
