+++ src/main/java/com/nhnacademy/boostorenginx/controller/WelcomeCouponController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.boostorenginx.controller;
 
 import com.nhnacademy.boostorenginx.dto.welcome.WelComeCouponRequestDto;
-import com.nhnacademy.boostorenginx.service.WelcomeCouponService;
+import com.nhnacademy.boostorenginx.service.impl.WelcomeCouponServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,11 +14,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/coupons")
 @RestController
 public class WelcomeCouponController {
-    private final WelcomeCouponService welcomeCouponService;
+    private final WelcomeCouponServiceImpl welcomeCouponServiceImpl;
 
     @PostMapping("/welcome")
     public ResponseEntity<String> issueWelcomeCoupon(@RequestBody WelComeCouponRequestDto requestDto) {
-        welcomeCouponService.issueWelcomeCoupon(requestDto);
+        welcomeCouponServiceImpl.issueWelcomeCoupon(requestDto);
         return ResponseEntity.ok("웰컴 쿠폰이 성공적으로 발급되었습니다");
     }
 }
