+++ src/main/java/com/nhnacademy/boostorenginx/service/MemberCouponService.java
@@ -9,9 +9,9 @@ public interface MemberCouponService {
 
     void useMemberCoupon(MemberCouponUseRequestDto dto); // 회원이 쿠폰을 사용
 
-    Page<MemberCouponResponseDto> getMemberCouponsByCouponId(MemberCouponFindByCouponIdRequestDto requestDto); // 쿠폰의 ID로 회원쿠폰 조회
+    Page<MemberCouponResponseDto> getMemberCouponsByCouponId(MemberCouponFindByCouponIdRequestDto requestDto); // 쿠폰 ID 로 회원쿠폰 목록 조회
 
-    Page<MemberCouponGetResponseDto> getMemberCouponsByMemberId(MemberCouponFindByMemberIdRequestDto requestDto); // 회원 ID에 해당되는 회원쿠폰 관계들을 모두 조회
+    Page<MemberCouponGetResponseDto> getMemberCouponsByMemberId(MemberCouponFindByMemberIdRequestDto requestDto); // 회원 ID 에 연관된 회원쿠폰 목록 조회
 
     Page<MemberCouponGetResponseDto> getUnusedMemberCouponsByMemberId(MemberCouponFindByMemberIdRequestDto requestDto); // 회원 ID 와 Status 로 회원쿠폰 조회
 }
